Ollie matched sub_8007858

+++ include/sprite_4.h
@@ -1,7 +1,7 @@
 #ifndef GUARD_SPRITE_4_H
 #define GUARD_SPRITE_4_H
 
-extern void sub_8007858(u8 param0, u8 param1, u8 param2, u32 param3, u16 param4);
+extern void sub_8007858(u8 param0, u8 param1, u8 param2, u16 param3, u16 param4);
 extern void sub_80078D4(u8 bg, int_vcount minY, int_vcount maxY, u16 offsetEven, u16 offsetOdd);
 
 #endif
